+++ frontend/src/routes.tsx
@@ -1,5 +1,6 @@
 import { Navigate } from "react-router-dom";
 import AppLayout from "src/components/layouts/AppLayout";
+import ReceiveLayout from "src/components/layouts/ReceiveLayout";
 import SendLayout from "src/components/layouts/SendLayout";
 import SettingsLayout from "src/components/layouts/SettingsLayout";
 import TwoColumnFullScreenLayout from "src/components/layouts/TwoColumnFullScreenLayout";
@@ -63,6 +64,7 @@ import Receive from "src/screens/wallet/Receive";
 import Send from "src/screens/wallet/Send";
 import SignMessage from "src/screens/wallet/SignMessage";
 import WithdrawOnchainFunds from "src/screens/wallet/WithdrawOnchainFunds";
+import ReceiveInvoice from "src/screens/wallet/receive/ReceiveInvoice";
 import ConfirmPayment from "src/screens/wallet/send/ConfirmPayment";
 import LnurlPay from "src/screens/wallet/send/LnurlPay";
 import PaymentSuccess from "src/screens/wallet/send/PaymentSuccess";
@@ -99,8 +101,19 @@ const routes = [
           },
           {
             path: "receive",
-            element: <Receive />,
             handle: { crumb: () => "Receive" },
+            element: <ReceiveLayout />,
+            children: [
+              {
+                index: true,
+                element: <Receive />,
+              },
+              {
+                handle: { crumb: () => "Invoice" },
+                path: "invoice",
+                element: <ReceiveInvoice />,
+              },
+            ],
           },
           {
             path: "send",
